5 bash scripts from Gary to create time series
Also, a single python script that submits each of the five scripts to the slurm
queue on casper.

Note that I've modified Gary's original scripts to take case identifier (e.g.
003 or 004) and a single year as command line arguments. The python script sets
the default case to 004, but requires users to specify at least a single year.
The user can also specify specific scripts to run, and the default is to run
all five of them. There is also a "dry-run" option that doesn't actually call
sbatch.