CaseClass tracks file type opened by gen_dataset
There is now a way to query whether a specific year of a variable from dataset
came from time series or history files. This is probably only useful for the
compare_ts_and_hist notebooks, which have  been re-run. Note that for this
commit I re-ran the notebooks on cheyenne, which does not have access to the
time series data on campaign -- when casper is back up, I will re-run the
notebooks to actually do the comparison.