Add savefig support for histograms and time series
Note that I noticed a pretty big bug to fix in the next commit -- I'm not
accounting for isel_dict in the file name, so plotting the same variable at two
vertical levels will create a plot at the first level and then over-write it
with the plot at the second level. I'll fix that tomorrow.