Instead of isel_dict, metadata lists sel values
I replaced metadata['isel_dict'] with metadata['reduced_dims'] and instead of
including the index of the dim reduction, it lists the actual value (e.g.
instead of basins=0, it contains basins="Global")

I had also switched from ncar-jobqueue to dask-jobqueue in a previous commit,
but with ncar-jobqueue now recognizing the PBS nodes on casper I was able to
switch back.